move the rest of dbengine config to netdata-conf
