+++ src/daemon/config/netdata-conf-db.h
@@ -3,8 +3,22 @@
 #ifndef NETDATA_DAEMON_NETDATA_CONF_DBENGINE_H
 #define NETDATA_DAEMON_NETDATA_CONF_DBENGINE_H
 
-#include "netdata-conf.h"
+#include "libnetdata/libnetdata.h"
+
+extern bool dbengine_enabled;
+extern size_t storage_tiers;
+extern bool dbengine_use_direct_io;
+
+extern int default_rrd_update_every;
+extern int default_rrd_history_entries;
+extern int gap_when_lost_iterations_above;
+extern time_t rrdset_free_obsolete_time_s;
+
+size_t get_tier_grouping(size_t tier);
 
 void netdata_conf_section_db(void);
+void netdata_conf_dbengine_init(const char *hostname);
+
+#include "netdata-conf.h"
 
 #endif //NETDATA_DAEMON_NETDATA_CONF_DBENGINE_H
