+++ src/daemon/config/netdata-conf-global.h
@@ -3,9 +3,11 @@
 #ifndef NETDATA_NETDATA_CONF_GLOBAL_H
 #define NETDATA_NETDATA_CONF_GLOBAL_H
 
-#include "netdata-conf.h"
+#include "libnetdata/libnetdata.h"
 
 void netdata_conf_section_global(void);
 void netdata_conf_section_global_run_as_user(const char **user);
+
+#include "netdata-conf.h"
 
 #endif //NETDATA_NETDATA_CONF_GLOBAL_H
