+++ src/database/rrd.c
@@ -14,8 +14,6 @@
 int rrd_delete_unupdated_dimensions = 0;
 */
 
-int default_rrd_update_every = UPDATE_EVERY;
-int default_rrd_history_entries = RRD_DEFAULT_HISTORY_ENTRIES;
 #ifdef ENABLE_DBENGINE
 RRD_MEMORY_MODE default_rrd_memory_mode = RRD_MEMORY_MODE_DBENGINE;
 #else
